requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-community
langchain-text-splitters
numpy
faiss-cpu
langchain-openai
openai
tiktoken
rank_bm25
pandas
matplotlib
